% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\docType{data}
\name{categories}
\alias{categories}
\title{Marginal Revolution post categories}
\format{
Data frame with columns
\describe{
\item{id}{Post ID}
\item{category}{Post category}
}
}
\source{
\href{https://marginalrevolution.com}{Marginal Revolution}
}
\usage{
data(categories)
}
\description{
Data frame containing post ID-category pairs.
}
\examples{
categories

if (require("dplyr")) {
categories \%>\% count(category)
}

}
\keyword{datasets}
