% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Marginal Revolution post metadata}
\format{
Data frame with columns
\describe{
\item{path}{Post path}
\item{id}{Post ID}
\item{time}{Post publication time}
\item{title}{Post title}
\item{author}{Post author}
}
}
\source{
\href{https://marginalrevolution.com}{Marginal Revolution}
}
\usage{
data(metadata)
}
\description{
Data frame containing metadata on Marginal Revolution blog posts.
}
\examples{
metadata

if (require("dplyr")) {
metadata \%>\% count(author)
}

}
\keyword{datasets}
